move make_request to helpers
