--- openbb_sdk/sdk/provider/openbb_provider/utils.py
@@ -1,75 +1,0 @@
-"""Generic providers utils"""
-
-import random
-
-import requests
-
-
-def get_user_agent() -> str:
-    """Get a not very random user agent."""
-    user_agent_strings = [
-        "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.10; rv:86.1) Gecko/20100101 Firefox/86.1",
-        "Mozilla/5.0 (Windows NT 6.1; WOW64; rv:86.1) Gecko/20100101 Firefox/86.1",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.10; rv:82.1) Gecko/20100101 Firefox/82.1",
-        "Mozilla/5.0 (Macintosh; Intel Mac OS X 10.13; rv:86.0) Gecko/20100101 Firefox/86.0",
-        "Mozilla/5.0 (Windows NT 10.0; WOW64; rv:86.0) Gecko/20100101 Firefox/86.0",
-        "Mozilla/5.0 (Macintosh; U; Intel Mac OS X 10.10; rv:83.0) Gecko/20100101 Firefox/83.0",
-        "Mozilla/5.0 (Windows NT 6.1; WOW64; rv:84.0) Gecko/20100101 Firefox/84.0",
-    ]
-
-    return random.choice(user_agent_strings)  # nosec
-
-
-def make_request(
-    url: str, method: str = "GET", timeout: int = 10, **kwargs
-) -> requests.Response:
-    """Abstract helper to make requests from a url with potential headers and params.
-
-    Parameters
-    ----------
-    url : str
-        Url to make the request to
-    method : str, optional
-        HTTP method to use.  Can be "GET" or "POST", by default "GET"
-    timeout : int, optional
-        Timeout in seconds, by default 10.  Can be overwritten by user setting, request_timeout
-
-    Returns
-    -------
-    requests.Response
-        Request response object
-
-    Raises
-    ------
-    ValueError
-        If invalid method is passed
-    """
-    # We want to add a user agent to the request, so check if there are any headers
-    # If there are headers, check if there is a user agent, if not add one.
-    # Some requests seem to work only with a specific user agent, so we want to be able to override it.
-    headers = kwargs.pop("headers", {})
-    preferences = kwargs.pop("preferences", None)
-    if preferences and "request_timeout" in preferences:
-        timeout = preferences["request_timeout"] or timeout
-
-    if "User-Agent" not in headers:
-        headers["User-Agent"] = get_user_agent()
-
-    # Allow a custom session for caching, if desired
-    _session = kwargs.pop("session", None) or requests
-
-    if method.upper() == "GET":
-        return _session.get(
-            url,
-            headers=headers,
-            timeout=timeout,
-            **kwargs,
-        )
-    if method.upper() == "POST":
-        return _session.post(
-            url,
-            headers=headers,
-            timeout=timeout,
-            **kwargs,
-        )
-    raise ValueError("Method must be GET or POST")
